+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.13.0 - 2020-06-15
+#### New
+- New output : **Rocketchat**
+- New output : **Mattermost**
+
 # 2.12.3 - 2020-04-21
 #### Enhancement
 - Allow using Datadog EU site by specifying new variable **datadog.host**/**DATADOG_HOST** ([PR#59](https://github.com/falcosecurity/falcosidekick/pull/59) thanks to [@DrPhil](https://github.com/DrPhil))